chore: no need to checkin .DS_Store
The .DS_Store file is a system file used by macOS to store custom attributes of a folder, such as the position of icons or the choice of a background image. It's not relevant to the project code or functionality, and including it can lead to unnecessary clutter in your repository.

Signed-off-by: Rui Chen <rui@chenrui.dev>